Fix bad create_user call, set email on creation
Was calling create_user with UNUSABLE_PASSWORD as the email address,
and None as the password, which resulted in UNUSABLE_PASSWORD being used
anyway, but also resulted in the email being set to a fixed value, which
fails in DBs where email is constrained unique. Now passes a unique
initial email address based on the username (which will generally be
overwritten by CAS_USERINFO_CALLBACK whenever it matters).